fixing XD for exclusion segments

+++ CrossDevicer/src/main/scala/CrossDevicer.scala
@@ -520,7 +520,7 @@ object CrossDevicer {
       .load("/data/metadata/segment_exclusion.tsv")
       .select("_c0", "_c1", "_c3") // group, segment id, score
       .rdd
-      .map(x => (x(1).toString.toInt, (x(0).toString.toInt, x(2).toString.toInt))) // (Segment ID -> (Group, Score))
+      .map(x => (x(1).toString.toInt, (x(0).toString, x(2).toString.toInt))) // (Segment ID -> (Group, Score))
       .collect()
       .toArray
       .toMap
